Move experiment configs to own file
