+++ src/ert/run_models/ensemble_smoother.py
@@ -10,20 +10,17 @@ from ert.config import (
     PreExperimentFixtures,
 )
 from ert.ensemble_evaluator import EvaluatorServerConfig
+from ert.experiment_configs import EnsembleSmootherConfig
 from ert.run_models.initial_ensemble_run_model import (
     InitialEnsembleRunModel,
-    InitialEnsembleRunModelConfig,
 )
-from ert.run_models.update_run_model import UpdateRunModel, UpdateRunModelConfig
+from ert.run_models.update_run_model import UpdateRunModel
 from ert.trace import tracer
 
 from ..run_arg import create_run_arguments
-from .run_model import ErtRunError, ExperimentType
+from .run_model import ErtRunError
 
 logger = logging.getLogger(__name__)
-
-
-class EnsembleSmootherConfig(InitialEnsembleRunModelConfig, UpdateRunModelConfig): ...
 
 
 class EnsembleSmoother(InitialEnsembleRunModel, UpdateRunModel, EnsembleSmootherConfig):
@@ -85,7 +82,3 @@ class EnsembleSmoother(InitialEnsembleRunModel, UpdateRunModel, EnsembleSmoother
     @classmethod
     def description(cls) -> str:
         return "Sample parameters → evaluate → update → evaluate"
-
-    @classmethod
-    def _experiment_type(cls) -> ExperimentType:
-        return ExperimentType.ENSEMBLE_SMOOTHER
