+++ src/ert/run_models/evaluate_ensemble.py
@@ -1,24 +1,19 @@
 from __future__ import annotations
 
 import logging
-from typing import Any, ClassVar
+from typing import Any
 from uuid import UUID
 
 import numpy as np
 
 from ert.ensemble_evaluator import EvaluatorServerConfig
-from ert.storage.local_experiment import ExperimentType
+from ert.experiment_configs import EvaluateEnsembleConfig
 from ert.trace import tracer
 
 from ..run_arg import create_run_arguments
-from .run_model import RunModel, RunModelConfig
+from .run_model import RunModel
 
 logger = logging.getLogger(__name__)
-
-
-class EvaluateEnsembleConfig(RunModelConfig):
-    ensemble_id: str
-    supports_rerunning_failed_realizations: ClassVar[bool] = True
 
 
 class EvaluateEnsemble(RunModel, EvaluateEnsembleConfig):
@@ -70,7 +65,3 @@ class EvaluateEnsemble(RunModel, EvaluateEnsembleConfig):
     @classmethod
     def description(cls) -> str:
         return "Use existing parameters → evaluate"
-
-    @classmethod
-    def _experiment_type(cls) -> ExperimentType:
-        return ExperimentType.EVALUATE_ENSEMBLE
