+++ src/ert/run_models/run_model.py
@@ -19,12 +19,11 @@ from contextlib import contextmanager
 from dataclasses import dataclass
 from datetime import datetime
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, ClassVar, Protocol, cast
+from typing import TYPE_CHECKING, Any, Protocol, cast
 
 import numpy as np
 from pydantic import (
     PrivateAttr,
-    computed_field,
     field_validator,
 )
 from pydantic_core.core_schema import ValidationInfo
@@ -38,15 +37,12 @@ from _ert.events import (
 from ert.config import (
     ConfigValidationError,
     DesignMatrix,
-    ForwardModelStep,
     HookedWorkflowFixtures,
-    HookRuntime,
     ModelConfig,
     ParameterConfig,
     PostSimulationFixtures,
     PreSimulationFixtures,
     QueueConfig,
-    Workflow,
     create_workflow_fixtures_from_hooked,
 )
 from ert.config.queue_config import KnownQueueOptionsAdapter
@@ -63,6 +59,7 @@ from ert.ensemble_evaluator.state import (
     REALIZATION_STATE_FAILED,
     REALIZATION_STATE_FINISHED,
 )
+from ert.experiment_configs import RunModelConfig
 from ert.mode_definitions import MODULE_MODE
 from ert.runpaths import Runpaths
 from ert.storage import (
@@ -71,13 +68,11 @@ from ert.storage import (
     Storage,
     open_storage,
 )
-from ert.storage.local_experiment import ExperimentType
 from ert.trace import tracer
 from ert.utils import log_duration
 from ert.warnings import PostExperimentWarning, capture_specific_warning
 from ert.workflow_runner import WorkflowRunner
 
-from ..base_model_context import BaseModelWithContextSupport
 from ..run_arg import RunArg
 from ._create_run_path import create_run_path
 from .event import (
@@ -161,25 +156,6 @@ class RunModelAPI:
     start_simulations_thread: StartSimulationsThreadFn
     cancel: Callable[[], None]
     has_failed_realizations: Callable[[], bool]
-
-
-class RunModelConfig(BaseModelWithContextSupport):
-    storage_path: str
-    runpath_file: Path
-    user_config_file: Path
-    env_vars: dict[str, str]
-    env_pr_fm_step: dict[str, dict[str, Any]]
-    runpath_config: ModelConfig
-    queue_config: QueueConfig
-    forward_model_steps: list[ForwardModelStep]
-    substitutions: dict[str, str]
-    hooked_workflows: defaultdict[HookRuntime, list[Workflow]]
-    active_realizations: list[bool]
-    log_path: Path
-    random_seed: int
-    start_iteration: int = 0
-    minimum_required_realizations: int = 0
-    supports_rerunning_failed_realizations: ClassVar[bool] = False
 
 
 class RunModel(RunModelConfig, ABC):
@@ -328,14 +304,15 @@ class RunModel(RunModelConfig, ABC):
         """
         return None
 
-    @classmethod
-    @abstractmethod
-    def _experiment_type(cls) -> ExperimentType: ...
+    # @classmethod
+    # @abstractmethod
+    # def _experiment_type(cls) -> ExperimentType: ...
 
-    @computed_field(return_type=str)  # type: ignore[prop-decorator]
-    @property
-    def experiment_type(self) -> ExperimentType:
-        return self.__class__._experiment_type()
+    # @computed_field(return_type=str)  # type: ignore[prop-decorator]
+    # @property
+    # def experiment_type(self) -> ExperimentType:
+    #    self.experiment_type
+    #    return self.__class__._experiment_type()
 
     def send_event(self, event: StatusEvents) -> None:
         self._status_queue.put(event)
