+++ src/ert/run_models/single_test_run.py
@@ -1,17 +1,10 @@
 from __future__ import annotations
 
-from pydantic import Field
-
+from ert.experiment_configs import SingleTestRunConfig
+from ert.experiment_types import ExperimentType
 from ert.run_models import EnsembleExperiment
-from ert.run_models.ensemble_experiment import EnsembleExperimentConfig
-from ert.storage.local_experiment import ExperimentType
 
 SINGLE_TEST_RUN_GROUP = "Forward model evaluation"
-
-
-class SingleTestRunConfig(EnsembleExperimentConfig):
-    active_realizations: list[bool] = Field(default_factory=lambda: [True])
-    minimum_required_realizations: int = 1
 
 
 class SingleTestRun(EnsembleExperiment, SingleTestRunConfig):
