+++ src/ert/storage/local_experiment.py
@@ -33,6 +33,7 @@ from ert.config._create_observation_dataframes import (
     create_observation_dataframes,
 )
 from ert.config._observations import Observation
+from ert.experiment_types import ExperimentType
 
 from .mode import BaseMode, Mode, require_write
 
@@ -50,19 +51,6 @@ class ExperimentState(StrEnum):
     stopped = auto()
     failed = auto()
     never_run = auto()
-
-
-class ExperimentType(StrEnum):
-    UNDEFINED = "Undefined"
-    SINGLE_TEST_RUN = "Single Test Run"
-    ENSEMBLE_EXPERIMENT = "Ensemble Experiment"
-    EVALUATE_ENSEMBLE = "Evaluate Ensemble"
-    ES_MDA = "Multiple Data Assimilation"
-    ENSEMBLE_SMOOTHER = "Ensemble Smoother"
-    ENSEMBLE_INFORMATION_FILTER = "Ensemble Information Filter"
-    MANUAL_UPDATE = "Manual Update"
-    MANUAL = "Manual"
-    EVEREST = "Everest"
 
 
 class ExperimentStatus(BaseModel):
